Move more content to metadata.py
