+++ physionet-django/project/models.py
@@ -7,8 +7,7 @@ import logging
 
 from background_task import background
 from django.conf import settings
-from django.contrib.contenttypes.fields import GenericForeignKey, GenericRelation
-from django.contrib.contenttypes.models import ContentType
+from django.contrib.contenttypes.fields import GenericRelation
 from django.db import models, transaction
 from django.forms.utils import ErrorList
 from django.urls import reverse
@@ -62,126 +61,6 @@ def move_files_as_readonly(pid, make_zip):
 
     if make_zip:
         published_project.make_zip()
-
-class Topic(models.Model):
-    """
-    Topic information to tag ActiveProject/ArchivedProject
-    """
-    content_type = models.ForeignKey(ContentType, on_delete=models.CASCADE)
-    object_id = models.PositiveIntegerField()
-    project = GenericForeignKey('content_type', 'object_id')
-
-    description = models.CharField(max_length=50, validators=[validate_topic])
-
-    class Meta:
-        unique_together = (('description', 'content_type', 'object_id'),)
-
-    def __str__(self):
-        return self.description
-
-    def save(self, *args, **kwargs):
-        super().save(*args, **kwargs)
-        self.project.content_modified()
-
-    def delete(self, *args, **kwargs):
-        super().delete(*args, **kwargs)
-        self.project.content_modified()
-
-
-class PublishedTopic(models.Model):
-    """
-    Topic information to tag PublishedProject
-    """
-    projects = models.ManyToManyField('project.PublishedProject',
-        related_name='topics')
-    description = models.CharField(max_length=50, validators=[validate_topic])
-    project_count = models.PositiveIntegerField(default=0)
-
-    def __str__(self):
-        return self.description
-
-
-class Reference(models.Model):
-    """
-    Reference field for ActiveProject/ArchivedProject
-    """
-    content_type = models.ForeignKey(ContentType, on_delete=models.CASCADE)
-    object_id = models.PositiveIntegerField()
-    project = GenericForeignKey('content_type', 'object_id')
-
-    description = models.CharField(max_length=1000)
-
-    class Meta:
-        unique_together = (('description', 'content_type', 'object_id'),)
-
-    def __str__(self):
-        return self.description
-
-    def save(self, *args, **kwargs):
-        super().save(*args, **kwargs)
-        self.project.content_modified()
-
-    def delete(self, *args, **kwargs):
-        super().delete(*args, **kwargs)
-        self.project.content_modified()
-
-
-class PublishedReference(models.Model):
-    """
-    """
-    description = models.CharField(max_length=1000)
-    project = models.ForeignKey('project.PublishedProject',
-        related_name='references', on_delete=models.CASCADE)
-
-    class Meta:
-        unique_together = (('description', 'project'))
-
-
-class Contact(models.Model):
-    """
-    Contact for a PublishedProject
-    """
-    name = models.CharField(max_length=120)
-    affiliations = models.CharField(max_length=150)
-    email = models.EmailField(max_length=255)
-    project = models.OneToOneField('project.PublishedProject',
-        related_name='contact', on_delete=models.CASCADE)
-
-
-class BasePublication(models.Model):
-    """
-    Base model for the publication to cite when referencing the
-    resource
-    """
-    citation = models.CharField(max_length=1000)
-    url = models.URLField(blank=True, default='')
-
-    class Meta:
-        abstract = True
-
-class Publication(BasePublication):
-    """
-    Publication for ArchivedProject/ActiveProject
-    """
-    content_type = models.ForeignKey(ContentType, on_delete=models.CASCADE)
-    object_id = models.PositiveIntegerField()
-    project = GenericForeignKey('content_type', 'object_id')
-
-    def save(self, *args, **kwargs):
-        super().save(*args, **kwargs)
-        self.project.content_modified()
-
-    def delete(self, *args, **kwargs):
-        super().delete(*args, **kwargs)
-        self.project.content_modified()
-
-
-class PublishedPublication(BasePublication):
-    """
-    Publication for published project
-    """
-    project = models.ForeignKey('project.PublishedProject',
-        db_index=True, related_name='publications', on_delete=models.CASCADE)
 
 class UnpublishedProject(models.Model):
     """
